Example of how R and python h2o-3 processes can be started with context_path argument so they fit nicely behind a reverse proxy.
The R method is the correct way, but the python method is a kludge because the python
api doesn't support the context_path argument as part of h2o.init().  This is a bug.

(Note that sparkling water doesn't need any explicit arguments in the startup process
of each notebook because we achieve the same thing through the global spark
configuration file.)